+++ CHANGES.rst
@@ -3,6 +3,7 @@
 
 - Changed stylings in all topic list page and reduce font size for 11th posts and on.
 - Added posts abbreviation for post exceeding 500 characters in topic list page.
+- Added per-board posting rate limit.
 - Added temporary favicon.
 
 0.5.1